Inject the cache instance
In case a user (or a test) needs to or wants to use their own Cache
object, I have added the ability to inject the cache. The cache_minutes
and cache_hours parameters have been moved to **kwargs, since it does
not make sense to require those if a custom cache is being injected.

If the user does not inject a cache, a default one will be created for
them. The default cache stores the objects in the temporary directory.